ah-ha-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai # Allowing latest version to align with genai.Client() usage
python-dotenv
google-auth
google-auth-oauthlib
python-jose[cryptography]
google-cloud-firestore
passlib[bcrypt]
google-adk